+++ src/komm/_lossless_coding/__init__.py
@@ -1,11 +1,15 @@
+from .FanoCode import FanoCode
 from .FixedToVariableCode import FixedToVariableCode
 from .HuffmanCode import HuffmanCode
+from .ShannonCode import ShannonCode
 from .TunstallCode import TunstallCode
 from .VariableToFixedCode import VariableToFixedCode
 
 __all__ = [
+    "FanoCode",
     "FixedToVariableCode",
     "HuffmanCode",
+    "ShannonCode",
     "TunstallCode",
     "VariableToFixedCode",
 ]