+++ tests/lossless_coding/test_huffman_code.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 
 import komm
+from komm._util.information_theory import random_pmf
 
 
 def test_huffman_code_1():
@@ -81,12 +82,29 @@ def test_huffman_code_invalid_call():
         komm.HuffmanCode([0.5, 0.5], policy="unknown")  # type: ignore
 
 
-@pytest.mark.parametrize("source_cardinality", [2, 3, 4, 5, 6])
-@pytest.mark.parametrize("source_block_size", [1, 2])
+@pytest.mark.parametrize("source_cardinality", range(2, 7))
+@pytest.mark.parametrize("source_block_size", range(1, 4))
+@pytest.mark.parametrize("policy", ["high", "low"])
+def test_huffman_code_random_pmf(source_cardinality, source_block_size, policy):
+    for _ in range(10):
+        pmf = random_pmf(source_cardinality)
+        code = komm.HuffmanCode(pmf, source_block_size=source_block_size, policy=policy)
+        assert code.is_uniquely_decodable()
+        assert code.is_prefix_free()
+        assert code.kraft_parameter() <= 1
+        entropy = komm.entropy(pmf)
+        assert entropy <= code.rate(pmf) <= entropy + 1 / source_block_size
+        # Permute pmf and check if the rate is the same.
+        pmf1 = pmf[np.random.permutation(source_cardinality)]
+        code1 = komm.HuffmanCode(pmf1, source_block_size, policy)
+        np.testing.assert_almost_equal(code.rate(pmf), code1.rate(pmf1))
+
+
+@pytest.mark.parametrize("source_cardinality", range(2, 7))
+@pytest.mark.parametrize("source_block_size", range(1, 4))
 @pytest.mark.parametrize("policy", ["high", "low"])
 def test_huffman_code_encode_decode(source_cardinality, source_block_size, policy):
-    integers = np.random.randint(0, 100, source_cardinality)
-    pmf = integers / integers.sum()
+    pmf = random_pmf(source_cardinality)
     dms = komm.DiscreteMemorylessSource(pmf)
     code = komm.HuffmanCode(pmf, source_block_size=source_block_size, policy=policy)
     x = dms(1000 * source_block_size)